remove *str_copyn and *str_catn_s, replace str_find_char with Str::FindChar

+++ baseutils/Vec.h
@@ -181,7 +181,7 @@ public:
 
     void Append(const T* src, size_t size=-1)
     {
-        if (-1 == size)
+        if ((size_t)-1 == size)
             size = Len(src);
         T* dst = MakeSpaceAt(len, size);
         memcpy(dst, src, size * sizeof(T));
